Settings can now scroll, added screensaver time setting, added min/max text to settings, general clean up
