+++ src/nvstorage.h
@@ -14,7 +14,9 @@ typedef enum eNVItem {
 	REFLOW_MIN_FAN_SPEED,
 	REFLOW_BAKE_SETPOINT_H,
 	REFLOW_BAKE_SETPOINT_L,
+
 	SCREENSAVER_ACTIVE,
+
 	NVITEM_NUM_ITEMS // Last value
 } NVItem_t;
