+++ src/spriteDefs.h
@@ -20,14 +20,22 @@ const uint16_t sprite9x16[] ={
 	0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, // <solid block>
 };
 
+// Pacman and ghost sprites - first 7 for , then 6 graphics for ghosts, with right-facing mirrored
 const uint16_t sprite16x16[] ={
+
+// Pacman - closed mouth ball, plus 2 more animation frames per direction - right facing graphic is mirrored in software
 	0b0000011111100000, 0b0001111111111000, 0b0011111111111100, 0b0111111111111110, 0b0111111111111110, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b0111111111111110, 0b0111111111111110, 0b0011111111111100, 0b0001111111111000, 0b0000011111100000,
+
 	0b0000011111100000, 0b0001111111111000, 0b0011111111111100, 0b0111111111111110, 0b0111111111111110, 0b1111111111111111, 0b1111111111111111, 0b1111111111111111, 0b1111111011111111, 0b1111111011111111, 0b1111110001111111, 0b0111110001111110, 0b0111100000111110, 0b0011100000111100, 0b0001000000011000, 0b0000000000000000,
 	0b0000011111100000, 0b0001111111111000, 0b0011111111111100, 0b0111111111111110, 0b0111111111111110, 0b1111111111111111, 0b1111111011111111, 0b1111110001111111, 0b1111100000111111, 0b1111000000011111, 0b1110000000001111, 0b0100000000000110, 0b0000000000000010, 0b0000000000000000, 0b0000000000000000, 0b0000000000000000,
+
 	0b0000011111100000, 0b0001111111111000, 0b0011111111111100, 0b0111111111111110, 0b0111111111111110, 0b1111111111111100, 0b1111111111110000, 0b1111111111000000, 0b1111111100000000, 0b1111111111000000, 0b1111111111110000, 0b0111111111111100, 0b0111111111111110, 0b0011111111111100, 0b0001111111111000, 0b0000011111100000,
 	0b0000011111100000, 0b0001111111111000, 0b0011111111110000, 0b0111111111100000, 0b0111111111000000, 0b1111111110000000, 0b1111111100000000, 0b1111111000000000, 0b1111110000000000, 0b1111111000000000, 0b1111111100000000, 0b0111111110000000, 0b0111111111000000, 0b0011111111100000, 0b0001111111110000, 0b0000011111100000,
+
 	0b0000011111100000, 0b0001111111111000, 0b0011111111111100, 0b0111111111111110, 0b0011111111111110, 0b0000111111111111, 0b0000001111111111, 0b0000000011111111, 0b0000001111111111, 0b0000111111111111, 0b0011111111111111, 0b0111111111111110, 0b0111111111111110, 0b0011111111111100, 0b0001111111111000, 0b0000011111100000,
 	0b0000011111100000, 0b0000111111111000, 0b0000011111111100, 0b0000001111111110, 0b0000000111111110, 0b0000000011111111, 0b0000000001111111, 0b0000000000111111, 0b0000000001111111, 0b0000000011111111, 0b0000000111111111, 0b0000001111111110, 0b0000011111111110, 0b0000111111111100, 0b0001111111111000, 0b0000011111100000,
+
+// Ghosts - 3 directions each with 2 frames of animation, with right-facing mirrored
 	0b0000000000000000, 0b1111111111100000, 0b0111111111111000, 0b0011111111111100, 0b0111111110001110, 0b1111111100000110, 0b1111111101110111, 0b0011111110001111, 0b0011111111111111, 0b1111111110001111, 0b1111111100000110, 0b0111111101110110, 0b0011111110001100, 0b0111111111111000, 0b1111111111100000, 0b0000000000000000,
 	0b0000000000000000, 0b0111111111100000, 0b1111111111111000, 0b1111111111111100, 0b0111111110001110, 0b0011111100000110, 0b0111111101110111, 0b1111111110001111, 0b1111111111111111, 0b0111111110001111, 0b0011111100000110, 0b0111111101110110, 0b1111111110001100, 0b1111111111111000, 0b0111111111100000, 0b0000000000000000,
 	0b0000000000000000, 0b1111111111100000, 0b0111111111111000, 0b0011111111111100, 0b0111111110001110, 0b1111111100010110, 0b1111111100010111, 0b0011111110001111, 0b0011111111111111, 0b1111111110001111, 0b1111111100010110, 0b0111111100010110, 0b0011111110001100, 0b0111111111111000, 0b1111111111100000, 0b0000000000000000,
