+++ src/ui_extras.c
@@ -1,0 +1,307 @@
+/*
+ * ui_extras.c
+ *
+ *  Created on: 19 Feb 2019
+ *      Author: scott
+ */
+
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <string.h>
+#include "lcd.h"
+#include "rtc.h"
+#include "sensor.h"
+#include "reflow.h"
+#include "reflow_profiles.h"
+#include "keypad.h"
+#include "nvstorage.h"
+#include "setup.h"
+
+#include "ui_extras.h"
+
+spriteStruct sprite[TOTAL_SPRITES];
+
+extern uint8_t logobmp[];
+extern uint8_t stopbmp[];
+extern uint8_t selectbmp[];
+extern uint8_t editbmp[];
+extern uint8_t f3editbmp[];
+extern uint8_t graph2bmp[];
+
+uint8_t animTicker=0;
+uint8_t spriteAnimIX=0;
+uint8_t blinkCnt=0,blinkOn=0;
+
+int16_t totalReflowTicks=0;
+int8_t reflowDisplay=0;
+
+uint16_t screensaverTimeout=600;
+uint16_t screensaverCnt=0;
+
+
+void showHeader(char *s){
+	LCD_disp_str((uint8_t*)" ", 1, 0, 0, FONT6X6);
+	LCD_disp_str((uint8_t*)"<<<<<<<<", 8, 5-animTicker, 0, FONT6X6);
+
+	LCD_disp_str((uint8_t*)" ", 1, (20*6)+2, 0, FONT6X6);
+	LCD_disp_str((uint8_t*)">>>>>>>>", 8, (12*6)+2+animTicker, 0, FONT6X6);
+
+	uint8_t len = strlen(s);
+	LCD_disp_str((uint8_t*)s, len, LCD_ALIGN_CENTER(len), 6 * 0, FONT6X6);
+
+	if(++animTicker==6){
+		animTicker=0;
+	}
+}
+
+void showBar(uint16_t v,uint8_t y){
+	if(v>52)
+		v=52;
+	v=52-v;
+	uint8_t n,i;
+	for(n=0;n<6;n++){
+		for(i=0;i<v;i++){
+			LCD_ClearPixel(i+7-n,y+n);
+		}
+	}
+}
+
+void initSprites(void){
+
+	int16_t initA[5][7]={
+		{-15,24, 2, 0, 3, 3, 0},
+		{42,-15, 3, 1, 1, 2, 6},
+		{84,-15, 3, 1, 2, 2, 6},
+		{62,63,  1, 1, 1, 2, 6},
+		{102,63, 1, 1, 2, 2, 6},
+	};
+
+	for(uint8_t n=0;n<TOTAL_SPRITES;n++){
+		// dirs: 0=left,1=up,2=right,3=down
+		sprite[n].x=			initA[n][0];
+		sprite[n].y=			initA[n][1];
+		sprite[n].dir=			initA[n][2];
+		sprite[n].changeCnt=	(rand()%60)+20;
+		sprite[n].animFrame=	initA[n][3];
+		sprite[n].speed=		initA[n][4];
+		sprite[n].startFrame=	initA[n][3];
+		sprite[n].totalFrames=	initA[n][5];
+		sprite[n].baseFrame=	initA[n][6];
+		sprite[n].state=		0;	// 0=normal, 1=eaten
+	}
+
+}
+
+void drawSprites(void){
+	LCD_FB_Clear();
+
+	++spriteAnimIX;
+
+	for(uint8_t n=0;n<TOTAL_SPRITES;n++){
+
+		uint8_t animFrame=sprite[n].animFrame;
+		uint8_t spriteHorizFlip=0;
+
+		if(sprite[n].state==0){
+			if(spriteAnimIX&1){
+				if(++sprite[n].animFrame==sprite[n].totalFrames+sprite[n].startFrame){
+					sprite[n].animFrame=sprite[n].startFrame;
+				}
+			}
+
+			switch(sprite[n].dir){
+
+			case 0:	// left
+				sprite[n].x-=sprite[n].speed;
+				if(sprite[n].x<-15)
+					sprite[n].x+=128+16;
+				spriteHorizFlip=FLIP_HORIZONTAL;
+				break;
+
+			case 1: // up
+				sprite[n].y-=sprite[n].speed;
+				if(sprite[n].y<-15)
+					sprite[n].y+=64+16;
+				if(animFrame>0)
+					animFrame+=2;
+				break;
+
+			case 2:	// right
+				sprite[n].x+=sprite[n].speed;
+				if(sprite[n].x>127)
+					sprite[n].x-=128+16;
+				break;
+
+			case 3:	// down
+				sprite[n].y+=sprite[n].speed;
+				if(sprite[n].y>63)
+					sprite[n].y-=64+16;
+				if(animFrame>0)
+					animFrame+=4;
+				break;
+			}
+
+			if(--sprite[n].changeCnt<=0){
+				sprite[n].changeCnt=(rand()%(60/sprite[n].speed))+20;
+				if((rand()%2)==0){
+					--sprite[n].dir;
+				}else{
+					++sprite[n].dir;
+				}
+				if(sprite[n].dir<0)
+					sprite[n].dir=3;
+				else if(sprite[n].dir>3)
+					sprite[n].dir=0;
+			}
+
+			// Check for collision between ghost and pacman - pacman eats ghosts :)
+			if(n>0 && sprite[n].state==0){
+				if(!(sprite[n].x<sprite[0].x-10 || sprite[n].x>sprite[0].x+10 || sprite[n].y<sprite[0].y-10 || sprite[n].y>sprite[0].y+10)){
+					sprite[n].state=1;
+				}
+			}
+		}else{
+			animFrame=7;
+			if(--sprite[n].y<-10){
+				sprite[n].state=0;
+				switch(rand()%4){
+				case 0:
+					sprite[n].x=rand()%112;
+					sprite[n].y=-15;
+					sprite[n].dir=3;
+					break;
+				case 1:
+					sprite[n].x=rand()%112;
+					sprite[n].y=63;
+					sprite[n].dir=1;
+					break;
+				case 2:
+					sprite[n].y=rand()%48;
+					sprite[n].x=-15;
+					sprite[n].dir=2;
+					break;
+				case 3:
+					sprite[n].y=rand()%48;
+					sprite[n].x=127;
+					sprite[n].dir=0;
+					break;
+				}
+			}
+		}
+
+		LCD_DrawSprite(animFrame+sprite[n].baseFrame,sprite[n].x,sprite[n].y,SPRITE16X16|spriteHorizFlip);
+	}
+}
+
+void displayReflowScreen(uint32_t keyspressed, uint8_t modeChange){
+	uint32_t ticks = RTC_Read();
+	int8_t len=0;
+	char buf[22];
+
+	if(++blinkCnt>=5){
+		blinkOn=(blinkOn==0?1:0);
+		blinkCnt=0;
+	}
+
+	if (keyspressed & KEY_F1) {
+		--reflowDisplay;
+	}
+	if (keyspressed & KEY_F2) {
+		++reflowDisplay;
+	}
+
+	if(reflowDisplay>1){
+		reflowDisplay=0;
+	}else if(reflowDisplay<0){
+		reflowDisplay=1;
+	}
+
+	if(modeChange){
+		// Work out total number of ticks (seconds) for profile
+		uint8_t n;
+		for(n=0;n<NUMPROFILETEMPS-1;n++){
+			if(Reflow_GetSetpointAtIdx(n)==0 && Reflow_GetSetpointAtIdx(n+1)==0){
+				totalReflowTicks=n-1;
+				break;
+			}
+		}
+		if(totalReflowTicks<=0){
+			totalReflowTicks=n;
+		}
+		totalReflowTicks*=10;
+	}
+
+	int16_t diff=Reflow_GetSetpoint()-Reflow_GetActualTemp();
+
+	if(reflowDisplay==0){
+		Reflow_PlotDots();
+
+		LCD_BMPDisplay(stopbmp, 127 - 17, 0);
+
+
+		len = snprintf(buf, sizeof(buf), "SET %03u", Reflow_GetSetpoint());
+		LCD_disp_str((uint8_t*)buf, len, 15, 0, (diff>5 && blinkOn==1?FONT6X6|INVERT:FONT6X6));
+
+		len = snprintf(buf, sizeof(buf), "ACTUAL %03u", Reflow_GetActualTemp());
+
+		LCD_disp_str((uint8_t*)buf, len, 68, 0, (diff<-5 && blinkOn==1?FONT6X6|INVERT:FONT6X6));
+
+		len = snprintf(buf, sizeof(buf), "%03u", (unsigned int)ticks);
+		LCD_disp_str((uint8_t*)"RUN", 3, 110, 31, FONT6X6);
+		LCD_disp_str((uint8_t*)buf, len, 110, 37, FONT6X6);
+
+	}else if(reflowDisplay==1){
+		LCD_FB_Clear();
+		LCD_BMPDisplay(graph2bmp, 0, 0);
+
+		uint16_t v=(((13000*4)/200)*Reflow_GetSetpoint())/1000;	// use Sensor_GetTemp() for float
+		showBar(v,3);
+
+		v=(((13000*4)/200)*Reflow_GetActualTemp())/1000;
+		showBar(v,24);
+
+		v=((52000/totalReflowTicks)*(unsigned int)ticks)/1000;
+		showBar(52-v,45);
+
+		if(blinkOn==1 && diff>5){
+			LCD_DrawSprite(10,63,3,SPRITE9X16);
+			LCD_DrawSprite(10,73,3,SPRITE9X16);
+			LCD_DrawSprite(10,83,3,SPRITE9X16);
+			LCD_DrawSprite(10,96,3,SPRITE9X16);
+		}else{
+			LCD_drawBigNum(Reflow_GetSetpoint(),3, 63,3,SPRITE9X16|INVERT);
+			LCD_DrawSprite(0,96,3,SPRITE9X16|INVERT);
+		}
+
+		uint16_t t=(Sensor_GetTemp(TC_AVERAGE)*10);
+		if(blinkOn==1 && diff<-5){
+			LCD_DrawSprite(10,63,24,SPRITE9X16);
+			LCD_DrawSprite(10,73,24,SPRITE9X16);
+			LCD_DrawSprite(10,83,24,SPRITE9X16);
+			LCD_DrawSprite(10,96,24,SPRITE9X16);
+		}else{
+			LCD_drawBigNum(t/10,3, 63,24,SPRITE9X16|INVERT);
+			LCD_DrawSprite((int8_t)(t%10),96,24,SPRITE9X16|INVERT);
+		}
+
+		LCD_drawBigNum((totalReflowTicks-ticks)/60,2, 63,45,SPRITE9X16|INVERT);
+		LCD_drawBigNum((totalReflowTicks-ticks)%60,2, 86,45,SPRITE9X16|INVERT);
+
+	}
+}
+
+uint8_t timeForScreensaver(){
+	if(screensaverTimeout==0)	// disabled
+		return 0;
+	if(++screensaverCnt>=screensaverTimeout){
+		screensaverCnt=0;
+		return 1;
+	}
+	return 0;
+}
+
+void initScreensaverTimeout(){
+	screensaverCnt=0;
+	screensaverTimeout=NV_GetConfig(SCREENSAVER_ACTIVE)*600;	// Outside range of floats, so can't use the Setup_getValue method
+}
